// sdk/models/basic_CarrierFacility.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type basic_CarrierFacility = {
    address: string;
    city: string;
    code?: string;
    delivery_type: string;
    name: string;
    postal_code: string;
};