// sdk/models/basic_DeliveryDetails.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type basic_DeliveryDetails = {
    carrier: string;
    city: string;
    delivery_center_code: string;
    delivery_type: string;
    email: string;
    full_name: string;
    house_number: string;
    house_number_suffix: string;
    phone: string;
    postal_code: string;
    street: string;
};