// sdk/models/basic_PickUpDetails.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type basic_PickUpDetails = {
    carrier: string;
    city: string;
    delivery_type: string;
    email: string;
    full_name: string;
    house_number: string;
    house_number_suffix: string;
    phone: string;
    pick_up_center_code: string;
    postal_code: string;
    street: string;
};