// sdk/models/basic_ShippoAddress.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type basic_ShippoAddress = {
    city: string;
    country: string;
    full_name: string;
    phone: string;
    state: string;
    street_1: string;
    street_2?: string;
    zip_code: string;
};