// sdk/models/basic_ShippoParcel.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type basic_ShippoParcel = {
    distance_unit: string;
    height: number;
    length: number;
    mass_unit: string;
    shipment_date: string;
    weight: number;
    width: number;
};