// sdk/models/basic_ShippoRate.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type basic_ShippoRate = {
    amount: string;
    amount_local: string;
    arrives_by?: string;
    attributes: Array<string>;
    currency: string;
    currency_local: string;
    duration_terms: string;
    estimated_days: number;
    id: string;
    provider: string;
    provider_image: string;
};