// sdk/models/ShippingDetails.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ShippingDetailsAddress } from './ShippingDetailsAddress.ts';
export type ShippingDetails = {
    address: ShippingDetailsAddress;
    name: string;
    phone: string;
};